Update main.py
In this commit, the code checks if the `response['choices'][0]['text']` is not empty before sending the message. If the response is empty, it sends a message to the user saying that it is not able to send a response and asking them to try again later.
Also, you can add a try-except block around the code that makes the API call to handle any exception that could occur when making the API call.
Please let me know if you have any further questions or concerns.